parking_finder_project/requirements.txt
backports.ssl-match-hostname==3.5.0.1
cached-property==1.3.0
colorama==0.3.7
cycler==0.10.0
Django==1.10.6
django-cors-headers==2.0.2
django-extensions==1.7.8
django-webpack-loader==0.4.1
djangorestframework==3.6.2
docker==2.2.1
docker-compose==1.11.2
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
enum34==1.1.6
functools32==3.2.3.post2
image==1.5.5
ipaddress==1.0.18
jsonschema==2.6.0
matplotlib==2.0.0
nose==1.3.7
numpy==1.12.1
olefile==0.44
Pillow==4.0.0
psycopg2==2.7.1
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
requests==2.11.1
scikit-image==0.13.0
scipy==0.19.0
six==1.10.0
subprocess32==3.2.7
texttable==0.8.8
websocket-client==0.40.0
